+++ src/main/java/com/fighthub/dto/aluno/CriarAlunoRequest.java
@@ -1,4 +1,4 @@
-package com.fighthub.dto;
+package com.fighthub.dto.aluno;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
