+++ src/main/java/com/fighthub/dto/auth/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.fighthub.dto;
+package com.fighthub.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
